REF/TST: collect indexing tests by method (#37677)
