+++ pandas/tests/frame/indexing/test_indexing.py
@@ -110,21 +110,6 @@ class TestDataFrameIndexing:
         with pytest.raises(KeyError, match="not in index"):
             frame[idx]
 
-    def test_getitem_callable(self, float_frame):
-        # GH 12533
-        result = float_frame[lambda x: "A"]
-        expected = float_frame.loc[:, "A"]
-        tm.assert_series_equal(result, expected)
-
-        result = float_frame[lambda x: ["A", "B"]]
-        expected = float_frame.loc[:, ["A", "B"]]
-        tm.assert_frame_equal(result, float_frame.loc[:, ["A", "B"]])
-
-        df = float_frame[:3]
-        result = df[lambda x: [True, False, True]]
-        expected = float_frame.iloc[[0, 2], :]
-        tm.assert_frame_equal(result, expected)
-
     def test_setitem_list(self, float_frame):
 
         float_frame["E"] = "foo"
