+++ pandas/tests/series/indexing/test_datetime.py
@@ -57,23 +57,6 @@ def test_fancy_setitem():
     assert (s[48:54] == -3).all()
 
 
-def test_dti_reset_index_round_trip():
-    dti = date_range(start="1/1/2001", end="6/1/2001", freq="D")._with_freq(None)
-    d1 = DataFrame({"v": np.random.rand(len(dti))}, index=dti)
-    d2 = d1.reset_index()
-    assert d2.dtypes[0] == np.dtype("M8[ns]")
-    d3 = d2.set_index("index")
-    tm.assert_frame_equal(d1, d3, check_names=False)
-
-    # #2329
-    stamp = datetime(2012, 11, 22)
-    df = DataFrame([[stamp, 12.1]], columns=["Date", "Value"])
-    df = df.set_index("Date")
-
-    assert df.index[0] == stamp
-    assert df.reset_index()["Date"][0] == stamp
-
-
 def test_slicing_datetimes():
     # GH 7523
