+++ pandas/tests/series/indexing/test_numeric.py
@@ -86,16 +86,3 @@ def test_slice_float_get_set(datetime_series):
         datetime_series[4.5:10.0]
     with pytest.raises(TypeError, match=msg.format(key=r"4\.5")):
         datetime_series[4.5:10.0] = 0
-
-
-def test_slice_floats2():
-    s = Series(np.random.rand(10), index=np.arange(10, 20, dtype=float))
-
-    assert len(s.loc[12.0:]) == 8
-    assert len(s.loc[12.5:]) == 7
-
-    i = np.arange(10, 20, dtype=float)
-    i[2] = 12.2
-    s.index = i
-    assert len(s.loc[12.0:]) == 8
-    assert len(s.loc[12.5:]) == 7
